@@ -11,49 +11,81 @@ description: Vivarium Control App
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 2.0.0+1
+publish_to: none
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^0.5.2
-  firebase_auth: ^0.18.3
-  firebase_auth_web: ^0.3.2
-  cloud_firestore: ^0.14.3
-  google_sign_in: ^4.5.6
-  flutter_blue: ^0.7.3
-  flutter_bluetooth_serial: ^0.2.2
-  toast: ^0.1.5
-  fluttertoast: ^7.1.1
-  random_string: ^2.1.0
-  hive_flutter: ^0.3.1
-  hive: ^1.4.4+1
-  particle_setup: ^0.3.1
-  flutter_settings_screens: ^0.2.1
-  http: ^0.12.2
-  json_annotation: ^3.1.0
-  settings_ui: ^0.4.0
-  path: ^1.7.0
-  date_time_picker: ^1.1.0
-  charts_flutter: ^0.9.0
-  rflutter_alert: ^1.1.0
-  flutter_colorpicker: ^0.3.4
-  flutter_material_color_picker: ^1.0.5
-  flutter_staggered_grid_view: ^0.3.2
-  fl_animated_linechart: ^1.1.4
-  font_awesome_flutter: ^8.10.0
-  permission_handler: ^5.0.1+1
-  #flutter_ble_lib: ^2.3.0
-  flutter_reactive_ble: ^2.7.1
 
+  #Firebase
+  firebase_core: "^1.10.0"
+  firebase_auth: "^3.2.0"
+  firebase_messaging: ^11.1.0
+  firebase_database: ^8.1.0
+  firebase_storage: ^10.1.0
+
+  #notifications
+  flutter_local_notifications: ^9.1.2
+
+  #Permissions
+  permission_handler: ^8.3.0
+
+  #Provider
+  provider: ^6.0.1
+
+  #Google
+  google_sign_in: ^5.2.1
+
+  # Font & Icons
+  font_awesome_flutter: ^9.2.0
+
+  #Bluetooth
+  flutter_reactive_ble: ^3.1.0
+
+  #Bluetooth enabler
+  flutter_blue_elves: ^0.1.4
+
+  #Toast
+  fluttertoast: ^8.0.8
+
+  #Navigation bar
+  bottom_navy_bar: ^6.0.0
+
+  #Charts - null safety missing
+  fl_animated_linechart:
+    git:
+      url: https://github.com/JacobJust/fl_animated_linechart
+
+  #Grid view
+  flutter_staggered_grid_view: ^0.4.1
+
+  #Color picker
+  flex_color_picker: ^2.2.0
+
+  #Date Time picker
+  date_time_picker: ^2.1.0
+
+  #Confirm dialog
+  confirm_dialog: ^1.0.0
+
+  #Settings
+  app_settings: ^4.1.1
+
+  #Location request
+  location: ^4.3.0
+
+  #Check tank
+  flutter_background_service: ^0.1.5
+
+  #check internet connection
+  connectivity_plus: ^2.0.3
 
 dev_dependencies:
-  hive_generator: ^0.8.2
-  build_runner: ^1.10.4
-  pedantic: ^1.9.2
+#  lints: ^1.0.1
   flutter_test:
     sdk: flutter
 
@@ -72,8 +104,9 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+    - assets/background/main_full.jpg
+    - assets/icons/circle/aquarium_512.png
+
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
