@@ -4,35 +4,35 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class FullScreenView extends StatefulWidget {
-  final String address;
+  final String? address;
   final int param;
 
-  const FullScreenView({Key key, this.address, this.param}) : super(key: key);
+  const FullScreenView({Key? key, this.address, required this.param}) : super(key: key);
 
   @override
   _FullScreenViewState createState() => _FullScreenViewState();
 }
 
 class _FullScreenViewState extends State<FullScreenView> {
-  Timer _timer;
-  String _url;
-  int _param;
-  Image _image;
+  Timer? _timer;
+  late String _url;
+  late int _param;
+  late Image _image;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.black,
       body: GestureDetector(
+        onTap: () {
+          Navigator.pop(context);
+        },
         child: Center(
           child: Hero(
             tag: 'imageHero',
             child: _image
           ),
         ),
-        onTap: () {
-          Navigator.pop(context);
-        },
       ),
     );
   }
@@ -51,17 +51,16 @@ class _FullScreenViewState extends State<FullScreenView> {
 
   @override
   void dispose() {
-    _timer.cancel();
+    _timer!.cancel();
     super.dispose();
   }
 
   @override
   void didChangeDependencies() {
-    // TODO: implement didChangeDependencies
     super.didChangeDependencies();
     precacheImage(_image.image, context);
     if (_timer != null) {
-      _timer.cancel();
+      _timer!.cancel();
     }
 
     _timer = Timer.periodic(
