@@ -2,47 +2,78 @@ import 'package:flutter/material.dart';
 import 'package:vivarium_control_unit/utils/converters.dart';
 
 class OverviewCard extends StatelessWidget {
-  final String title;
+  final String? title;
   final List<String> additionalHeaders;
-  final Widget body;
+  final Widget? additionalHeader;
+  final Widget? body;
   final Color background;
+  final Color? titleBackground;
 
-  const OverviewCard({Key key,
-    this.title,
-    this.additionalHeaders = const [],
-    this.body,
-    this.background = const Color(0xffeeeeee)})
+  const OverviewCard(
+      {Key? key,
+      this.title,
+      this.additionalHeaders = const [],
+      this.body,
+      this.titleBackground,
+      this.background = Colors.transparent,
+      this.additionalHeader})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Card(
-        color: background,
-        elevation: 5,
-        child: Column(
-          children: [
-            Container(
-              color: Colors.cyan.shade200,
-              height: 23,
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(title),
-                  ...additionalHeaders.map((e) => Text(e)).toList()
-                ],
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Card(
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+          color: Colors.black12,
+          elevation: 2,
+          child: Column(
+            children: [
+              Container(
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.only(
+                        topLeft: Radius.circular(10),
+                        topRight: Radius.circular(10)),
+                    gradient: LinearGradient(colors: [
+                      Theme.of(context).accentColor,
+                      titleBackground ?? Theme.of(context).accentColor
+                    ], stops: [
+                      0.6,
+                      1
+                    ])),
+                // color: titleBackground ?? Theme.of(context).accentColor,
+                height: 30,
+                child: Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Text(title!, style: Theme.of(context).textTheme.headline5),
+                      if (additionalHeader != null) additionalHeader!,
+                      ...additionalHeaders
+                          .map((e) => Text(e,
+                              style: Theme.of(context).textTheme.headline6))
+                          .toList()
+                    ],
+                  ),
+                ),
               ),
-            ),
-            if (body != null) body
-          ],
-        ));
+              if (body != null)
+                Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: body,
+                )
+            ],
+          )),
+    );
   }
 }
 
 class TriggerCardBody extends StatelessWidget {
-
-  final Widget icon;
-  final int nextTriggerTime;
-  final Widget nextChange;
+  final Widget? icon;
+  final int? nextTriggerTime;
+  final Widget? nextChange;
 
   /// Decide whether the next trigger exists.
   ///
@@ -50,52 +81,56 @@ class TriggerCardBody extends StatelessWidget {
   final bool hasTrigger;
 
   const TriggerCardBody(
-      {Key key, this.icon, this.nextTriggerTime, this.nextChange, this.hasTrigger = true})
+      {Key? key,
+      this.icon,
+      this.nextTriggerTime,
+      this.nextChange,
+      this.hasTrigger = true})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return CardBody(
       icon: icon,
-      children: hasTrigger?[
-        Text('Next trigger: ${getTimeStringFromTime(nextTriggerTime)}'),
-        SizedBox(
-          height: 10,
-        ),
-        Row(
-          children: [Text('Next change:'), nextChange],
-        )
-      ]:[
-        Text('No triggers created')
-      ],
+      children: hasTrigger
+          ? [
+              Text('Next trigger: ${getTimeStringFromTime(nextTriggerTime)}'),
+              SizedBox(
+                height: 10,
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [Text('Next change:'), nextChange!],
+              )
+            ]
+          : [Text('No triggers created')],
     );
   }
 }
 
 class CardBody extends StatelessWidget {
-  final Widget icon;
-  final List<Widget> children;
+  final Widget? icon;
+  final List<Widget>? children;
 
-  const CardBody({Key key, this.icon, this.children}) : super(key: key);
+  const CardBody({Key? key, this.icon, this.children}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Expanded(
-        child: Container(
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Container(child: icon),
-              Padding(
-                  padding: const EdgeInsets.only(left: 8.0),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: children,
-                  ))
-            ],
-          ),
-        ));
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        Container(child: icon),
+        Padding(
+            padding: const EdgeInsets.only(left: 8.0),
+            child: FittedBox(
+              fit: BoxFit.contain,
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: children!,
+              ),
+            ))
+      ],
+    );
   }
 }
-
